Restructured the repo
