+++ .gitignore
@@ -1,1 +1,5 @@
 /bin/
+.idea/*
+.classpath*
+.project*
+*.iml
