+++ Sorts/BogoSort.java
@@ -1,8 +1,6 @@
-package sort;
+package Sorts;
 
 import java.util.Random;
-
-import static sort.SortUtils.*;
 
 
 /**
@@ -19,7 +17,7 @@ public class BogoSort implements SortAlgorithm {
 
     private static <T extends Comparable<T>> boolean isSorted(T array[]){
         for(int i = 0; i<array.length - 1; i++){
-            if(less(array[i + 1], array[i])) return false;
+            if(SortUtils.less(array[i + 1], array[i])) return false;
         }
         return true;
     }
@@ -30,7 +28,7 @@ public class BogoSort implements SortAlgorithm {
 
         for (int i = 0; i < array.length; i++) {
             int randomIndex = i + random.nextInt(length - i);
-            swap(array, randomIndex, i);
+            SortUtils.swap(array, randomIndex, i);
         }
     }
 
@@ -49,11 +47,11 @@ public class BogoSort implements SortAlgorithm {
         BogoSort bogoSort = new BogoSort();
 
         // print a sorted array
-        print(bogoSort.sort(integers));
+        SortUtils.print(bogoSort.sort(integers));
 
         // String Input
         String[] strings = {"c", "a", "e", "b","d"};
 
-        print(bogoSort.sort(strings));
+        SortUtils.print(bogoSort.sort(strings));
     }
 }
