+++ Sorts/SortUtils.java
@@ -1,4 +1,4 @@
-package sort;
+package Sorts;
 
 import java.util.Arrays;
 import java.util.List;
